requirements.txt
kagglehub==0.3.4
numpy==2.1.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.8.1.78
pandas==1.5.3
Pillow==11.0.0
streamlit==1.26.0
torch==2.0.1
torchvision==0.15.2
~pencv_python_headless==4.8.0.76
